+++ src/main/java/soma/edupigateway/EdupiGatewayApplication.java
@@ -1,4 +1,4 @@
-package soma.edupi.gateway;
+package soma.edupigateway;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
